requirements.txt
langchain
pinecone-client
streamlit
unstructured
python-dotenv
openai